Revert "[NFC] Move 'InterModuleDependencyGraph' out-of-date analysis into an extension on the graph itself"
This reverts commit b588e0d0c21fb5754608ea4e799007b7fddf749e.
