+++ Sources/SwiftDriver/IncrementalCompilation/FirstWaveComputer.swift
@@ -138,6 +138,75 @@ extension IncrementalCompilationState.FirstWaveComputer {
             mandatoryJobsInOrder: mandatoryJobsInOrder)
   }
 
+  /// We must determine if any of the module dependencies require re-compilation
+  /// Since we know that a prior dependency graph was not completely up-to-date,
+  /// there must be at least *some* dependencies that require being re-built.
+  ///
+  /// If a dependency is deemed as requiring a re-build, then every module
+  /// between it and the root (source module being built by this driver
+  /// instance) must also be re-built.
+  private func computeInvalidatedModuleDependencies(on moduleDependencyGraph: InterModuleDependencyGraph)
+  throws -> Set<ModuleDependencyId> {
+    let mainModuleInfo = moduleDependencyGraph.mainModule
+    var modulesRequiringRebuild: Set<ModuleDependencyId> = []
+    var visited: Set<ModuleDependencyId> = []
+    // Scan from the main module's dependencies to avoid reporting
+    // the main module itself in the results.
+    for dependencyId in mainModuleInfo.directDependencies ?? [] {
+      try outOfDateModuleScan(on: moduleDependencyGraph, from: dependencyId, 
+                              visited: &visited, modulesRequiringRebuild: &modulesRequiringRebuild)
+    }
+
+    reporter?.reportExplicitDependencyReBuildSet(Array(modulesRequiringRebuild))
+    return modulesRequiringRebuild
+  }
+
+  /// Perform a postorder DFS to locate modules which are out-of-date with respect
+  /// to their inputs. Upon encountering such a module, add it to the set of invalidated
+  /// modules, along with the path from the root to this module.
+  private func outOfDateModuleScan(on moduleDependencyGraph: InterModuleDependencyGraph,
+                                   from moduleId: ModuleDependencyId,
+                                   visited: inout Set<ModuleDependencyId>,
+                                   modulesRequiringRebuild: inout Set<ModuleDependencyId>) throws {
+    let moduleInfo = try moduleDependencyGraph.moduleInfo(of: moduleId)
+    // Visit the module's dependencies
+    var hasOutOfDateModuleDependency = false
+    var mostRecentlyUpdatedDependencyOutput: TimePoint = .zero
+    for dependencyId in moduleInfo.directDependencies ?? [] {
+      // If we have not already visited this module, recurse.
+      if !visited.contains(dependencyId) {
+        try outOfDateModuleScan(on: moduleDependencyGraph, from: dependencyId,
+                                visited: &visited, modulesRequiringRebuild: &modulesRequiringRebuild)
+      }
+      // Even if we're not revisiting a dependency, we must check if it's already known to be out of date.
+      hasOutOfDateModuleDependency = hasOutOfDateModuleDependency || modulesRequiringRebuild.contains(dependencyId)
+
+      // Keep track of dependencies' output file time stamp to determine if it is newer than the current module.
+      if let depOutputTimeStamp = try? fileSystem.lastModificationTime(for: VirtualPath.lookup(moduleDependencyGraph.moduleInfo(of: dependencyId).modulePath.path)),
+         depOutputTimeStamp > mostRecentlyUpdatedDependencyOutput {
+        mostRecentlyUpdatedDependencyOutput = depOutputTimeStamp
+      }
+    }
+
+    if hasOutOfDateModuleDependency {
+      reporter?.reportExplicitDependencyWillBeReBuilt(moduleId.moduleNameForDiagnostic, reason: "Invalidated by downstream dependency")
+      modulesRequiringRebuild.insert(moduleId)
+    } else if try !IncrementalCompilationState.IncrementalDependencyAndInputSetup.verifyModuleDependencyUpToDate(moduleID: moduleId, moduleInfo: moduleInfo,
+                                                                                                                 fileSystem: fileSystem, reporter: reporter) {
+      reporter?.reportExplicitDependencyWillBeReBuilt(moduleId.moduleNameForDiagnostic, reason: "Out-of-date")
+      modulesRequiringRebuild.insert(moduleId)
+    } else if let outputModTime = try? fileSystem.lastModificationTime(for: VirtualPath.lookup(moduleInfo.modulePath.path)),
+              outputModTime < mostRecentlyUpdatedDependencyOutput {
+      // If a prior variant of this module dependnecy exists, and is older than any of its direct or transitive
+      // module dependency outputs, it must also be re-built.
+      reporter?.reportExplicitDependencyWillBeReBuilt(moduleId.moduleNameForDiagnostic, reason: "Has newer module dependency inputs")
+      modulesRequiringRebuild.insert(moduleId)
+    }
+
+    // Now that we've determined if this module must be rebuilt, mark it as visited.
+    visited.insert(moduleId)
+  }
+
   /// In an explicit module build, filter out dependency module pre-compilation tasks
   /// for modules up-to-date from a prior compile.
   private func computeMandatoryBeforeCompilesJobs() throws -> [Job] {
@@ -155,7 +224,7 @@ extension IncrementalCompilationState.FirstWaveComputer {
 
     // Determine which module pre-build jobs must be re-run
     let modulesRequiringReBuild =
-      try moduleDependencyGraph.computeInvalidatedModuleDependencies(fileSystem: fileSystem, reporter: reporter)
+        try computeInvalidatedModuleDependencies(on: moduleDependencyGraph)
 
     // Filter the `.generatePCM` and `.compileModuleFromInterface` jobs for 
     // modules which do *not* need re-building.
